fix: move trackers preview to main package
