+++ extension-manifest-v3/package.json
@@ -4,8 +4,9 @@
   "version": "10.4.5",
   "type": "module",
   "scripts": {
-    "download-engines": "node scripts/download-engines/index.js",
-    "download-whotracksme-bloomfilter": "node scripts/download-whotracksme-bloomfilter.js",
+    "download-engines": "node scripts/download-engines.js",
+    "download-wtm-bloomfilter": "node scripts/download-wtm-bloomfilter.js",
+    "download-wtm-stats": "node scripts/download-wtm-stats.js",
     "build": "node scripts/build.js",
     "start": "npm run build -- --watch",
     "licenses": "license-report --config=../.license-report-config.json > dist/licenses.html",
@@ -37,11 +38,10 @@
     "@cliqz/adblocker-webextension": "^1.33.2",
     "@cliqz/adblocker-webextension-cosmetics": "^1.33.2",
     "@duckduckgo/autoconsent": "^10.15.0",
-    "@ghostery/trackers-preview": "^1.0.0",
     "@ghostery/ui": "^1.0.0",
     "@github/relative-time-element": "^4.3.0",
-    "@whotracksme/reporting": "^5.1.21",
     "@sentry/browser": "^8.26.0",
+    "@whotracksme/reporting": "^5.1.21",
     "hybrids": "^9.1.1",
     "idb": "^7.1.1",
     "jwt-decode": "^4.0.0",
