+++ extension-manifest-v3/scripts/build.js
@@ -1,3 +1,14 @@
+/**
+ * Ghostery Browser Extension
+ * https://www.ghostery.com/
+ *
+ * Copyright 2017-present Ghostery GmbH. All rights reserved.
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0
+ */
+
 import { resolve, dirname, join } from 'path';
 import { readFileSync, writeFileSync, readdirSync } from 'fs';
 import { exec, execSync } from 'child_process';
@@ -54,7 +65,8 @@ if (argv.staging) {
   execSync('npm run download-engines', { stdio: 'inherit' });
 }
 
-execSync('npm run download-whotracksme-bloomfilter', { stdio: 'inherit' });
+execSync('npm run download-wtm-bloomfilter', { stdio: 'inherit' });
+execSync('npm run download-wtm-stats', { stdio: 'inherit' });
 
 const config = {
   logLevel: argv.silent ? 'silent' : undefined,
