+++ extension-manifest-v3/scripts/download-wtm-bloomfilter.js
