+++ extension-manifest-v3/scripts/download-wtm-stats.js
@@ -1,9 +1,20 @@
+/**
+ * Ghostery Browser Extension
+ * https://www.ghostery.com/
+ *
+ * Copyright 2017-present Ghostery GmbH. All rights reserved.
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0
+ */
+
 import { writeFileSync } from 'fs';
 import { URL, fileURLToPath } from 'url';
 
 const DATA_URL = 'https://whotracks.me/data/trackers-preview.json';
 const OUTPUT_FILE = fileURLToPath(
-  new URL('../src/background/trackers-preview-generated.js', import.meta.url),
+  new URL('../src/rule_resources/wtm-stats.js', import.meta.url),
 );
 
 const data = await fetch(DATA_URL).then((res) => {
