+++ extension-manifest-v3/src/content_scripts/trackers-preview.css
