--- packages/trackers-preview/.eslintignore
@@ -1,2 +1,0 @@
-/src/background/trackers-preview-generated.js
-/scripts/**/*.js
