--- packages/trackers-preview/.gitignore
@@ -1,1 +1,0 @@
-src/background/trackers-preview-generated.js
