--- packages/trackers-preview/package.json
@@ -1,23 +1,0 @@
-{
-  "name": "@ghostery/trackers-preview",
-  "version": "1.0.0",
-  "private": true,
-  "type": "module",
-  "author": "Ghostery GmbH",
-  "license": "MPL-2.0",
-  "scripts": {
-    "postinstall": "node scripts/download.js",
-    "lint": "eslint ./src",
-    "test": "npm run lint"
-  },
-  "exports": {
-    "./background": "./src/background/index.js",
-    "./content_scripts": "./src/content_scripts/index.js",
-    "./content_scripts/styles.css": "./src/content_scripts/styles.css",
-    "./page_scripts": "./src/page_scripts/index.js"
-  },
-  "dependencies": {
-    "@ghostery/ui": "^1.0.0",
-    "tldts-experimental": "^6.0.19"
-  }
-}
