+++ packages/ui/package.json
@@ -7,7 +7,6 @@
   "exports": {
     "./onboarding": "./src/modules/onboarding/index.js",
     "./panel": "./src/modules/panel/index.js",
-    "./trackers-preview": "./src/modules/trackers-preview/index.js",
     "./youtube": "./src/modules/youtube/index.js",
     "./youtube/wall": "./src/modules/youtube/wall.js",
     "./categories": "./src/utils/categories.js",
